Split Imageable trait into Imageable and Uploadable
